Lint the AutoGraph code
Add a separate .flake8.ag config that is used in a separate command when
running the make lint-python command.

Add a .style.yapf file mostly matching AutoGraph.

Relax the requirements to 100 lines in width.
Sprinkle the code with flake8's noqa directives where pylint was
previously used.
Fix a lot of new line issues that were introduced during rework.
Disable few checks that are commonly used patterns in AG test
and module code.
Fix formatting in few cases to not disable the whole classes of problems
as the alternative would be the noqa, and we are already changing the
code.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>