Remove `Operator` inheritance from `VideoDecoderBase` (#4508)
This PR changes the VideoDecoderBase not to inherit from Operator. There is no reason to retain this inheritance. VideoDecoderBase relies only on FramesDecoder and includes it by composition. Furthermore, this allows operators that actually inherit from VideoDecoderBase to inherit from other flavours of Operator, without the Diabolic Diamond Of Death.

Signed-off-by: szalpal <mszolucha@nvidia.com>