Extracting `InputOperator` from `ExternalSource` (#4505)
* Introducing InputOperator base class.
* Extracting common parts from ExternalSource to InputOperator
* Renaming external_source configuration structures.
* Moving ValidateInputs from enqueuing (SetDataSource) to operator run (SetupImpl).
* Adjusting external_source tests for new place for ValidateInputs.
* Removing external_source.cu - there's no CUDA code to be compiled there.

With the VideoInput there is a new category of operators emerging in DALI - Input Operators. The difference from Readers is that the Input Operators inject the data from memory to DALI pipeline. On the contrary - Readers read the data from disk.

ExternalSource has already been an Input Operator, but was never named this way. This PR extracts the common parts of InputOperators from ExternalSource.

InputOperator's API contains the following:

    1. SetDataSource - this function already has been present in ExternalSource. The Pipeline (or any other entity) uses it to enqueue the data.
    2. ForwardCurrentData - this is a new function. It's purpose is to populate given TensorList with the data that is currently on the top of the queue.

All the ExternalSource features (no-copy etc...) are forwarded to the InputOperators.

Lastly, several minor adjustments had to be made, such as structs renaming.

Signed-off-by: szalpal <mszolucha@nvidia.com>