Utilities and refactoring pre-`VideoInput` operator (#4513)
* Moving `caching_list.h` to builtin operators directory.
* Adding `InputOperator` documentation.
* Relaxing enforcing uniform batch size for InputOperators.
* Allowing feed_input to work with all InputOperators (instead of ExternalSource only).
* Introducing `DecodeFrames` in `VideoDecoderBase`.
* Reverting template method pattern in InputOperator's `SetupImpl`.

During the development of `VideoInput` it tuned out, that the template method pattern for InputOperator `SetupImpl` function does not suffice the behaviour requirements. The main reason was that `HandleDataAvailability` function has been called every Operator iteration, while `VideoInput` mustn't check it that often. On the contrary - `VideoInput` may work on the data fed a few iterations ago. Since `HandleDataAvailability` checks, if the data is enqueued - it might be a false result. Therefore, `HandleDataAvailability` shall be called by every subclass of `InputOperator`, what is stated in the documentation.

The `VideoDecoderBase` needed an additional method to decode parts of the video injected into `FramesDecoder`. This function received also a `padding` feature, required by `VideoInput` operator.

Lastly, the uniform-batch-size-requirement for the `InputOperators` had to be relaxed, since these operators actually create a batch. Therefore, they can decide not to create a uniform batch across iteration and it is perfectly fine behaviour.

Signed-off-by: szalpal <mszolucha@nvidia.com>