`VideoInput` operator for the CPU (#4519)
* Implementing `VideoInput` for CPU with test suite.
* Adding `inputs` tests to the `python-readers-decoders` suite.
* Relaxing `feed_input` check, to check the presence of the callback only in the `external_source`, not all InputOperators.
* Polishing `feed_input` documentation.
* Relaxing the Uniform Input/Output Batch Size check, not to cover the `VideoInput` operator.
* Exclude `VideoInput` from testing for variable batch size.
* Add CPU-only `VideoInput` test.
* Expose `InputOperator::ForwardCurrentData` function.

Introducing new test category - `inputs` - requires adding those to the test suite. Not to multiply existing suites, added those to the `readers-decoders` suite. Did not rename the `reader-decoder` to `readers-decoders-input`, not to make the name too long. Additionaly, such rename would carry the need of modifying CI scripts, which I wanted to avoid.

`feed_input` has been checking for the presence of the callback. The reason of this check is that the `external_source` operator has 2 way of specifying the input - callback and `feed_input`. These modes are mutually exclusive. Since in past no other operator could use `feed_input`, this was OK. Now it is changed to check only when `external_source` is in use.

The VideoInput is an operator, that might return various batch sizes throughout operation. While max_batch_size is constant throughout the pipeline lifetime, the VideoInput might return the partial or incomplete batch at the end of the video file. This happens, when the number of frames in the video is insufficient to fill the batch completely. For this reason, the check for a uniform batch size had to be relaxed.


Signed-off-by: Michał Szołucha <mszolucha@nvidia.com>