Operator trace mechanism (#4564)
TLDR:
* Introducing daliCreatePipeline2 function in C API.
* Introducing functions to use Operator Traces feature in the C API.
* Introduced a notion of an iteration_id.
* Adding an Operator Name to the Workspace.
* Introduced an IterationData.
* Added PassthroughOperatorWithTrace also to dali_operator_test binary.
* Added tests for operator traces.


This PR adds an Operator Trace mechanism. Until now, user had no inside to what happens in the particular operators - he has only seen whatever makes it to the output of the pipeline. The Operator Trace mechanism allows user to communicate with specific operators within the pipeline.

The operator traces are mainly required for the VideoInput operator. When using the VideoInput from the C API, user needs to know, when the VideoInput has consumed the data provided by external source. Using operator traces, user will be able to provide the data_id (this feature is the next step of this PR) and use it to identify the status of the input data in the VideoInput operator.

Operator traces are currently exposed only in C API. The entry points to the feature are the following functions:

int daliHasOperatorTrace(daliPipelineHandle *pipe_handle, const char *operator_name, const char *trace_name);
const char * daliGetOperatorTrace(daliPipelineHandle *pipe_handle, const char *operator_name, const char *trace_name);

Please refer to their documentation for more details.


Operator traces are kept in a new data structure called IterationData. The point of the operator traces is to keep the data for the whole iteration - maintaining it across the stages. Therefore this new data structure has been introduced. The executor keeps a queue of these IterationData, to properly handle the Prefetching.

While working on the operator traces, it became apparent, that we need an identifier of the iteration, that couples the workspaces across the stages. From now on, the Executor tracks the iteration ID. The iteration ID is used to access the IterationData corresponding to a given iteration. Also, to properly handle the OperatorTraceMap, the Workspace needs to know, which operator does it belong to. Therefore, an information about the Operator instance has been added to the Workspace.

To make the tests easier, PassthroughOperatorWithTrace has been added to the dali_operator_test binary. It's used only for testing.

Existing daliCreatePipeline function in the C API did not allow to create a Pipeline, that executes within a SimpleExecutor (i.e. the one that has exec_async=False and exec_pipelined=False). Since such Pipeline has been required for the test of the Operator Traces, daliCreatePipeline2 function has been created.

Signed-off-by: Michał Szołucha <mszolucha@nvidia.com>