`MixedBackend` support for `InputOperator` (#4603)
* Add In/Out type logic to the InputOperator.
* Add ForwardCurrentData specializations for MixedBackend.
* Refactor SetExternalInput routines and add MixedBackend handling.
* Add IdentityInput test-only operator.
* Add InputOperator<MixedBackend> tests.


So far we never had a `MixedBackend` input operator. None of: `ExternalSourceCpu`, `ExternalSourceGpu` and `VideoInputCpu` had to inherit from `InputOperator<MixedBackend>`. With `VideoInputGpu` the situation changes, since this operator is actually a mixed operator. Therefore the support for `MixedBackend` operators in the `InputOperator` had to be added.

The mixed input operator has somehow different meaning than the other two. Both CPU and GPU have a uniform device across the execution, while Mixed operator changes the device in the process. In other words, the following pattern occurs:

OpBackend   InpB OutB
CPU         CPU  CPU
MXD         CPU  GPU
GPU         GPU  GPU


Let's take an example of the operator, that subclasses `InputOperator`:
1. `MyOp<CPUBackend>`. This part is already handled by the `InputOperator`. Since operator takes the CPU input and returns a CPU output, the "intermediate" data (i.e. the data that is acquired using InputOperator API `ForwardCurrentData`) will also have a CPUBackend.
2. `MyOp<GPUBackend>`. This part is also handled already. Since the operator takes the GPU input and returns a GPU output, the intermediate data has to be on the GPU.
3. `MyOp<MixedOperator>`. This part is what the PR targets.

The input operator grabs the actual input from the top of the queue (i.e. `CachingList` in the implementation). Since it has a CPU input and returns a GPU output, we do not know what backend of the data the operator needs to proceed. It might be a CPU data (which I believe will be the more common case), when e.g. the operator need to conduct Huffman CPU-based decoding and then proceed with JPEG decoding on the GPU. However, maybe the operator does not need CPU data, but it needs to be a Mixed operator for any other reason - it this case, the operator would like `ForwardCurrentData` to return a tensor already with GPU backend.

For these reasons, there are 2 overloads of the `ForwardCurrentData`. One is for CPU and one for GPU target tensor. The former part would be more common and the latter part is just a convenience function, making the code a lot simpler. `ForwardCurrentData<Mixed>(Tensor<CPU>)` would behave in the same way as its CPU sibling - copy or swap depending on the pageability. In case of `ForwardCurrentData<Mixed>(Tensor<GPU>)` the data on the top of the queue is always a CPU one (since MixedOperator has a CPU input). Therefore the InputOperator needs to copy this data to the GPU.

For the CPU part, the operator has to create its own ThreadPool, since MixedOperators are not supplied with one automatically.

Signed-off-by: Michał Szołucha <mszolucha@nvidia.com>