`DataId` mechanism for `fn.inputs.video` operator (#4584)
* Add `data_id_map_t` to the C API `daliPipelineHandle`.
* Add setting the `next_output_data_id` trace in `VideoInput` operator.
* Switch loading the input data by `VideoInput` to end of `Run`.
* Minor `VideoInput` operator refactoring.
* Add tests for `VideoInput` operator with `next_output_data_id` trace.
* Add `data_id_` to the ExternalSource operator.
* Change the type inside by `CachingList` in `InputOperator` to type carrying also a `data_id`.
* Add `data_id` handling to the `InputOperator`.
* Add `data_id` handling to the `Pipeline`.
* Add a function, that removes an operator trace in the `Workspace`.

Overview:

To fully utilize the `VideoInput` operator, some new functionality to the API has to be added. Currently there's no other way of using `VideoInput` operator, than via `feed_input` function in Python or it's corresponding `daliSetExternalInput*(...)` functions in the C API. In case user does not know the total number of frames in the video beforehand, it's impossible to know how many time he may run `daliRun()` (or `pipeline.run()` in Python) on a given input sample.

To solve this problem (currently in C API only, this PR does not offer such possibility in Python) this PR introduces `DataId` mechanism. The point of this feature is to set the ID of the input data, that is fed to the `VideoInput` operator (actually, to any input operator, e.g. `external_source` too). Using the operator traces mechanism introduced in #4564 , `fn.inputs.video` after each iteration provides information, what will be the ID of the next input used by the decoder (or no ID, if it's unknown/no data in queue exists).

The API consists of only function. Please refer to its documentation to find more details.
daliSetExternalInputDataId(daliPipelineHandle *pipe_handle, const char *operator_name, const char *data_id);

Details:

The `data_id` has been introduces to the InputOperators. This action stipulated some changes in the existing input operators - external_source and video_input.

The DataId has to be carried around together with the data the ID refers to. Therefore change of the type, which holds the data in the InputOperator (actually in a CachingList in the InputOperator) was necessary. There have been some other possibilities of carrying the data_id (e.g. extending TensorList with it), but we decided to go with something targeted to the feature instead of introducing generic solution which might not be needed in the long run.

While implementing the `DataId` it turned out, that the OperatorTraces API wasn't full - it did not contain the function which would clear the trace of a given operator. Such function has been added in this PR.

The DataId feature introduces a change in how the VideoInput shall process data. Since it returns the data_id of the next input, it needs to check at the end of the Run, if it can and should load the next input. Otherwise, the situation when user feeds the VideoInput with multiple input samples before calling `Run` won't be handled properly.

Since the DataId from the perspective of the C API is handled similarly to the BatchSize, according structures has been introduced.


Signed-off-by: Michał Szołucha <mszolucha@nvidia.com>