Fix classification of argument input-only operators in AutoGraph (#4618)
Currently, an operator without inputs is considered a generator and is
registered in the global scope - it will always produce full batch,
and each use of such operator will insert appropriate split nodes 
- as there is no way of signaling the requested batch size from
the pipeline definition level. Some generators do not support variable
batch size at all.
Some operators can have 0 positional inputs while having some argument
inputs. Now they would be classified as generator and relegated
to the global scope.

Include detection of argument inputs when registering the outputs
of the operator - if the operator has any input or *argument input*,
it can detect the batch size of the input and infer the size
of the output.
This makes it usable in an `if` scope and doesn't require placing such
operators in global scope. Placing them in the global scope would result
in unnecessary splits to be inserted, even though the produced
batch size is already correct for given scope.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>