Fix @autoserialize problem with unknown module (#4628)
* Add exception handling to `@autoserialize` decorator.
* Add a test.

When autograph has been introduced to DALI, we've faced the problem with autoserialization. Turns out, that autograph in runtime swaps the modules using six - probably to maintain compatibility between Python2 and Python3. Therefore, when traversing the module tree in search for @autoserialize decorator, inspect has not been able to find some modules and was prompting an error.

The solution is that if given module is unavailable, it is not traversed any more. DALI won't be able to retrieve @autoserialize from there anyhow, so no harm done.

Signed-off-by: szalpal <mszolucha@nvidia.com>