`VideoInput<MixedBackend>` (#4613)
* Add a `VideoInput` operator for `mixed` device.
* Extend C++ and Python tests of the `VideoInput` to cover` mixed` device.
* Add `PadFrameCreator`.
* Refactor portions of `VideoInput` code - move implementation to header file.
* Add conditional compilation support (BUILD_NVDEC).
* Add a `ThreadPool` to `VideoInput` and `affine` argument to control it.
* Add awkward helper-structs `VideoInputCpu` and `VideoInputMixed`.

Finally, after many preliminary PRs, the time came to introduce GPU-based `VideoInput` operator. The purpose of this operator is to take an encoded Video file and gradually return it decoded in portions. It's capabilities are uniform with `VideoInputCpu` operator, therefore it's unnecessary to list all of them here.

The GPU-based `VideoInput` operator uses NVDEC (specialized NVIDIA hardware) for decoding videos. Since it takes a CPU input, it is actually a Mixed operator. The output is returned on the GPU.

Since MixedOperators do not have their own ThreadPool, one had to be supplied. Therefore, `affine` argument has been added to the `VideoInput` operator, to control the behaviour of the specialized ThreadPool.

Lastly, there is an inconsistency between APIs of `Operator<XPUBackend>` and `Operator<MixedBackend>` specialization. The `XPUBackend` specializations have a `RunImpl` function, which every specific operator shall override. On the other hand, the `MixedBackend` specialization does not have this function. It has a `Run` function instead. This leads to obvious problem when trying to write a operator for both backends. Unfortunately, this can't be resolved using SFINAE (`enable_if` for instance), because the `Run` and `RunImpl` are a virtual functions, which can't work with SFINAE. To avoid high code duplication, helper-struct have been introduces - `VideoInputCpu` and `VideoInputMixed`. Also, `ViRunImpl` function has been introduce for the sole purpose of calling it in a proper `Run...` override explicitly defined in these structs.

The implementation had to be moved to header file, to support conditional compilation. To properly compile (or not) the NVDEC-powered part, VideoInput had to be split into two separate compilation units.


Signed-off-by: szalpal <mszolucha@nvidia.com>