Extend input API utilities to support input operators (#4642)
* Make the functions: layout(), ndim(), dtype(), num_inputs() support all input operators.

This PR is a next one in the effort of introducing input operators. We had some API utilities that supported only external source. It was hardcoded. Now, since external_source is also an input operator, these utilities can and should support all input operators.

There is a minor change in the implementation. Since we shouldn't explore the operators by name now, I assumed that these utilities may be called provided the graph has been built. Therefore I can use dynamic casting to the InputOperator base class to determine, whether an operator in an input operator or not.

Signed-off-by: szalpal <mszolucha@nvidia.com>