Naive histogram custom operator example/template (#4615)
This PR adds an example of a custom operator. This is a fully-functioning example, that is ready to run and test in Python. The instructions are included in the README file.

The purpose of this contribution is to provide a template for an easy start when writing custom operators in DALI.

In this example, we're using a rather naive implementation of CUDA kernel computing an grayscale image histogram. Since the purpose of this is not to show how to write CUDA kernels, it's good enough.

Signed-off-by: szalpal <mszolucha@nvidia.com>