Add lazy `and` and `or`, and not lazy `not` support (#4629)
Add support for logical expressions when `conditionals_enabled=True`.
AutoGraph allows for the detection and overloading of logical expressions.

The `not` expression is executed eagerly as an operator - it's directly inserted in the AutoGraph overload.
For `and` and `or` AutoGraph captures the right argument of the logical expression as lambdas. We can utilize the same mechanism as in the `if` statement, and evaluate it lazily in accordance with proper Python semantics. 
The implementation is based on simplified form of the `if` statement. As we don't have an actual `and` and `or` operators, a special operator that validates both sides of the input is inserted in the binary logical expressions: `fn._conditional.validate_logical`.

All the logical expressions and `if` conditions are restricted to 0-d boolean inputs placed on CPU. The `boolean` part is important, as Python semantics for logical expression: https://docs.python.org/3/reference/expressions.html#boolean-operations require to evaluate the truthness (or call `__bool__`) of the inputs, and in the case of `and` and `or`, return the unchanged inputs. 

In cases like batches of: `[1, 0]` or `[0.0, 1.0]` it would mean that samplewise `or` results in `[1, 1.0]` - a batch of mixed types. If we restrict the inputs to just `booleans` we don't hit that problem.

The same error checking API as in `_conditional.validate_logical` is additionally added to check the inputs of the `if` statement condition and `not` operator to provide better and more consistent error messages between conditional use-cases. 
In `if` it is implemented by notifying the `split` operator that it is used as `if` statement implementation and it should apply the stricter checks. There is no need to do it in `merge`, as `split` is executed first.

All those operator land as builtin in a dedicated `conditional` directory, where the `split` and `merge` were moved for consistency.

The followup would be to include those expressions as a part of the tutorial/documentation.
There is another followup planned to lift some restrictions of boolean input type where it is possible without introducing the mentioned problems (for `if` condition and `not`).

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>