Augment exception handling with ImportError (#4681)
The exception handling in autoserialize guards from traversing the module, that might not exist. That sounds strange, but it happens, that a 3rd party library swaps version of a dependency just because (e.g. six). While such thing happens, one of two exceptions might be raised - ModuleNotFoundError and ImportError. We're adding the latter just in case.

Signed-off-by: szalpal <mszolucha@nvidia.com>