Reduce memory footprint of conditional execution
Keep only the currently required memory for operators producing outputs
within a conditional scope.

The executor detects if there were onditional statements used withing
the graph. In that case processing a partial batch is assumed to be an
execution of operator within a conditional scope.
For such operators their outputs are cleared of reserved memory before
run (and new allocation for that iteration) - this means that we keep
only the currently processed batch size instead of the biggest one we
saw.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>