Add support for more types in `if` predicate and `not` expression
Evaluate scalar numerical types treating them as truthy and falsy
values.
Utilize regular promotion from numeric types to boolean.
We can do it safely as `not` produces a new value and the predicate
in `if` is not related to the output.
`and` and `or` expression stay limited to `bool` inputs to not
risk creating scenarios where batches of mixed types can arise.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>