DALI Pipeline inputs as named arguments to `Pipeline.run()` (#4712)
Currently, providing inputs to DALI Pipeline is highly cumbersome. We have a couple of ways:

    1. Using Readers,
    2. Using external_source operator and its argument source,
    3. Using an input operator (e.g. external_source) and feed_input function.

All of these are quite problematic. This PR attempts to tackle this with allowing yet another way of providing inputs.

My proposal is to augment pipe.run() function with **kwargs in Python. Every such argument, provided that it has a corresponding input operator in DALI pipeline, can accept any data assigned to the name of such input operator.

The change is added run .run() only, since I figured it's always possible to add it to the schedule_run too, but removing it will be problematic due to backwards compatibility issues.

Signed-off-by: Michał Szołucha <mszolucha@nvidia.com>