Add `"depleted"` operator trace (#4794)
After introducing repeat_last option in the ExternalSource (#4775), an additional operator trace is needed to properly handle this option in DALI Backend. This PR adds the "depleted" trace. This trace is available for every InputOperator. By depleted == "true" operator informs, that it does not have sufficient resources to be ran again.

When repeat_last==True in the external source, it's obvious that it will never deplete, once the data has been provided to the operator.

* Add `depleted` operator trace.
* Handle `depleted` operator trace in VideoInput and ExternalSource operators.
* Add tests for `depleted` operator trace.
* Minor renames to better fit additional tests.

---------

Signed-off-by: Michał Szołucha <mszolucha@nvidia.com>