Support complex data structures as return values in conditionals
It was assumed, that the output of a helper function in case of
conditional code is plain tuple of DataNode.

Now the return values can be built out of any complex nestings of tuples
and dictionaries (they must still match between branches).

The Tree library is used to apply the splitting (and constant promotion)
for branch outputs and merging the if/else paths for aribitrarily nested
DataNodes instead of working only with tuples or single elements.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>