Add JAX iterator decorator (#5050)
It adds decorator API for JAX iterators. With this PR it is possible to write code:

@data_iterator(...):
def pipeline_def(...):
    ...
    return output


# This is an iterator returning JAX compatible outputs
iterator = pipeline_def(...)
Additional information:
There are 2 iterator for JAX DALIGenericIterator and DALIGenericPeekableIterator. Decorators were added for both.

To make sure that API does not diverge there are tests to make sure that decorators have the same args as the objective way of creating the iterators.

Decorators can be used both in declarative way:

@data_iterator(...)
def function(...):
    ....
and functional way:

data_iterator(function)(...)
Both are tested.
---------

Signed-off-by: Albert Wolant <awolant@nvidia.com>