Generalize `__module__` handling and hide private modules from Ops doc
After refactoring, the documentation for ops API started listing all the
internall modules. Modules that start with `_` are no longer visible
in Sphinx doc.

Moving the custom operator classes to the dedicated submodules casued
the docs to render the path to the implementation module and not the
one we want to present in the API - due to the `__module__` change.

As the module with the original implementation is still useful, a
generalized way of adjusting the `__module__` was introduced, keeping
the original information in the operator under _impl_module`.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>