Fix the check for dynamically generated number of outputs
The documentation for HasOutputFunction suggests that it allows to check
if operator calculates the number of outputs dynamically or if it is set
statically.

We have two entities:
OutputFn - calculates the number of outputs dynamically based on OpSpec
AdditionalOutputsFn - calculates dynamically based on OpSpec additional
outputs.

If any of those is present, number of outputs is not know statically.

HasOutputFn allowed to detect only the first case, now it will detect
any of the two.
This check is only used in Python when generating signatures.

This additionally fixes incorrect return type annotation for some
operators that indicate dynamic number of outputs using
AdditionalOutputsFn.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>