Expose Tensors types stubs in nvidia.dali.tensors module (#5153)
Base for the Python Tensor and TensorList types interface files,
generated using stubgen (for `nvidia.dali.backend_impl` module)
and inserted as `nvidia.dali.tensors`.

As those types are generated in the backend, this is needed for
tools like intellsense and mypy (working on static information)
to work.

C-based types (or types not guaranteed to be present, like np.array), 
are removed from signatures and left as a comment or replaced with Any.

Additionally, the `__module__` for Tensor types is adjusted to
`nvidia.dali.tensors` on the backend level. It allows pybind to generate
the documentation pointing to the desired module, rather than
the `backend_impl` one. 
We still have stubs which allow to find the Tensor types in the
`backend` and `backend_impl` modules, keeping the backward
compatibility.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>