Add `pmap` compatibility for JAX `data_iterator` (#5185)
Description:
This PR adds new syntax to data_iterator: new argument called devices. It can be used to configure the iterator to return output that is compatible with pmaped functions in JAX.

Additional information:
JAX has to ways of spreading the workload to multiple GPUs: automatic parallelization and pmap function. data_iterator supported automatic parallelization with sharding argument. For pmap support user had to manually create multiple pipelines and pass them to DALIGenericIterator. In this PR I added a way to use data_iterator with pmap.

With this introduction, I reworked tutorial for multigpu for DALI and JAX to reflect this change. While at it, I edited this tutorial to better reflect desired priority: automatic parallelization first, pmap second.

Signed-off-by: Albert Wolant <awolant@nvidia.com>