Introduce TensorLike to signatures utilizing Array Interface (#5179)
The generated signatures now include the TensorLike inputs when there is a possibility
to pass a DataNode.

There are two new type aliases introduced in the `_typing` module. 
* `TensorLikeIn` is intended to describe anything that can be passed as a constant to input. 
   This include arrays/tensors of accepted types, sequences of Python scalar numbers
   and the scalar numbers themselves (int and float). We don't have type constraints until
   the operator runs and validates the inputs, so we need to list all possibilities.
* `TensorLikeArg` is intended for argument inputs. As we already describe Sequences
   and scalar there (with a specific types as well), we keep this alias constrained
   to ArrayInterface.

There are two now protocol classes:
https://mypy.readthedocs.io/en/stable/protocols.html#simple-user-defined-protocols

They describe classes implementing:
* NumPy Array Interface (it has a special method `__array__`) and 
* Numba CUDA Array Interface (it has a property `__cuda_array_interface__`).

In the case of constants we actually handle only the NumPy Array Interface
 - and even this is an approximation. Our implementation handles only
PyTorch and MxNet tensors and NumPy arrays, using the `.flatten().to_list()` sequence.
Still, those classes do implement the `__array__()` method (even if only to throw
an exception when they are GPU-placed), and this is the closest and most generic
approximation I can come up with. 

The IDEs are also able to look up the docstrings for those aliases
and definitions when navigating the interface file.


Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>