Unify positional input name handling in docs and signatures (#5217)
Extract the common name processing and use the same naming format for
positional-only input parameter names for both generated docstrings and
the generated signatures.

The documentation generated names as `input0`, `input1`, ... (or just `input`). 
In the case of signatures, we used `__input_0` - the double underscore
at the start is the convention to indicate argument as positional only.
Based on the experiments, it is needed by some IDEs to not suggest setting
the name as a keyword parameter. Also, in case of named inputs, 
without the `__` there might be clash between them and the keywords
otherwise (for example `shape` in slice). 

Formatting is unified to the latter form of `__input_0`, `__input` in case there
is only one.

Adjust the way that the signatures are generated, so they list all mandatory inputs:
* signature has a cutoff point for too many inputs: 
   https://github.com/NVIDIA/DALI/blob/main/dali/python/nvidia/dali/ops/_signatures.py#L80
* Previously, if there were any kind of inputs (mandatory or optional, above 10)
   we just represented all of them as `*input`.
* Documentation lists all the mandatory inputs, and adds a generic: 
   `input[XX..YY]` for optional ones. 
* Now, the signatures are generated in the form:
   `def foo(__input_0, __input_1, ..., __input_9, /, *inputs)` - we will list all
   mandatory inputs, and hide the optional ones behind `VAR_POSITIONAL`
   only if it would be over threshold, to list all of them.
* DALI doesn't have an op with more than 3 mandatory inputs. 

This should match closer between the docs and the signatures.

Cleanup some todos on the way.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>