Allow source to be a decorated function or method (#5268)
This changes how the external source counts the number of arguments that the source callback uses.
Instead of relying on the `__code__.co_argcount` we use `inspect.signature`.
The reason is that a decorator can (and usually does) propagate the signature, but due to the use of
generic `*args, **kwargs` parameters it will hide the original argcount of the original function that is wrapped.

Such simple decorator is added as a test case, before this change the following code won't work.

This fixes the potential error of DALI assuming that there are mandatory arguments to the function,
that doesn't expect them (or vice-versa) and invoking the callback with (or without) passing the source_info. 
The error would point to the callback invocation and would not indicate the decorators at all.

Now we can sensibly error out on:
* var-args
* var-kwargs
* kwarg-only argument

There is no need to differentiate between positional only and named argument, if we first detect
any disallowed constructs. Note, that sources like `foo(arg, *args)` were already disallowed, but the reported arg count was something like `-1` or `-2` by mistake rather than by design.

Next we can validate the correct number of arguments.

The detection of callbacks was done inside an exception handler, and throwing another exception
from there would chain the exception, so the construction of argument description is moved to line below.


Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>