Enable AG-transformed code to show user code in exception
Autograph transforms all function invocations into converted_call
statement. When an exception is raised in such AG-transformed code,
every converted_call step will capture the exception and attach
additional metadata pointing to the original code. As we traverse only
the converted call, we produce call stack that is equivalent to the code
before the transformation.

The change in pipeline from to_graph to convert enables the final step -
attaching the equivalent call stack of user code to be added to the
produced exception.

This makes errors that happen during tracing the pipeline definition
more readable to the user by appending the "in user code <user
callstack>" message.

As the processing of pipeline definition arguments doesn't work with
decorators a WAR is used for now. The issue should be resolved in a
separate commit.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>