Reimplement parts of tf_stack with pure Python. Enable stack remapping
The solution to stack-trace manipulation in TF was optimized by moving
parts of the processing to C++. As we have limited needs for producing
traces, it is sufficient to replace them with regulart Python
constructs. This steps reenables the construction of StackTraceMapper
and Filter in the AG-converted code.

It will enable next steps:
* we will be capturing origin stack trace for every operator defintion
* in conditional mode we will remap it back to the user code from the
  AG-produced one.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>