Refactor TFRecordReader to use python_op_factory as base class
As _TFRecordReaderImpl is used twice with different schema
parametrization, it is now generated in a function, inheriting from the
same base class as all other ops.

This reduces the custom code to minimum, allowing the base class to
handle all properties and some common arguments.

The argument handling is moved to base class by updating the kwargs
before invoking the base class.
When a tfrec.Feature is encountered in the Python argument serialization
layer, it is again processed by the tfrec.Feature constructor (as we do
for other types, normalizing numbers with int() or float()).
For this purpose a copy constructor is added and exposed in Pybind.
An alternative would be to change the conversion to an identity
function - that would keep the error in the spec.AddArg rather than
moving it to constructor parameter matching.

Validation code is added to the operator, to check for matching type
with better error message.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>