Move module and op name partitioning to backend
The "__" delimiter splitting of schema name into a module path and
operator name is moved to backend.

Python APIs are switched over to query the schema.

ops.Compose is removed from registry of operators - it is mainly used to
generate documentation and stubs, Compose is not really an operator but
a special Python wrapper. It will be a standalone entity in the ops
module and the documentation for it is now listed separately.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>