Add origin stack trace capture for DALI operators (#5302)
Add capture of origin stack trace for DALI operators.
The `extract_stack` collects all frames from the first frame of pipeline definition to the frame
with operator invocation (the outermost user API).
In regular mode no further processing is needed.
For code that was transformed by AutoGraph, the contents of captured frames are remapped
back to the user code, filtered out of the _autograph and _conditionals modules (that contain
internal DALI implementation). 
Due to how autograph introduces additional frames (for example by implementing if statements
with additional function calls), we remove repeated occurrences of the same function, keeping
only the last one. AutoGraph's entry point to a function call is used to detect such regions.

Stack traces are collected in OperatorInstance, the outermost layer of the API adds an argument
denoting the current stack level, so we can compute how many frames we need to skip related
to DALI internals when capturing the stack.

The collected stack is added as hidden arguments to OpSpec. This is backward-compatible
for serialized pipelines, as well as allows for disabling and enabling the feature. 

As a default the feature is disabled and only available via hidden API.

C++ test operators are extended to work as a loadable plugin based on dummy operator test. 
It allows to implement tests for the new feature without extending regular DALI operators.

Follow-up to this PR will utilize this information for providing better error messages - pointing
to the origin of error in DALI pipeline definition.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>